Add files via upload
his project is a simple yet effective tool for encrypting and decrypting images using Python and the OpenCV library. It allows users to securely encode their images and messages, ensuring that sensitive information remains private. The core feature of this project is the pixel shuffling technique, which rearranges the pixels of the original image based on a user-defined seed key, making the encrypted image unrecognizable.